+++ task-manager/app/_layout.tsx
@@ -1,5 +1,12 @@
 import { Stack } from "expo-router";
 
 export default function RootLayout() {
-  return <Stack />;
+  return (
+  <Stack>
+      <Stack.Screen
+        name="index"
+        options={{ title: "Task List" }}
+      />
+    </Stack>
+  )
 }
